trace-cmd-virt-server: Add MSG_MAX to denote most supported cmd
In case we need to extend the number of types of messages, we do not want to
use FINMETA as a stopping point. The message types are going to be hard
coded to support various versions of the server. The MSG numbers can not be
changed once we release a version of trace-cmd with virt-server.

Add a new MSG_MAX that denotes the range of available commands.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>