trace-cmd msg: Move the saved closing fd to the caller
Instead of having a static variable that gets set to whatever file
descriptor is passed to tracecmd_msg_finish_sending_metadata() to be closed
with tracecmd_send_close_msg(), have that saved by the caller. As I can
imagine lots of nasty bugs happening later on with this side effect fd
saving. If the caller controls what file descriptor gets closed, it will be
more transparent for later development.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>