trace-cmd msg: Remove tracecmd_msg_send_and_wait_for_msg() helper function
The helper function tracecmd_msg_send_and_wait_for_msg() was used for very
specific purposes and not very robust. There was two users and those users
should open code the calls as they did not even use the helper function the
same way, which made the helper function rather pointless. It also prevented
checking the received message to see if it was what the caller really
wanted.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>