trace-cmd agent: Add agent to connect to host
Add a agent command that allows the guest to create a server that connects
with a trusted host that it will take commands from and start tracing with,
passing the data back to the host.

This currently only implements the connection of the guest to the host, but
does nothing more. It allows the host to register the guest, and also the
option of -A is added to list to show the listing of the registered guests.
But again, this only implements connecting a guest and listing them. More
needs to be done to implement executing the guest.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>